lesson-2-hosted-maf/foundry-agent/requirements.txt
# Microsoft Agent Framework - Agente de Mercado Financeiro
# Pinned versions para estabilidade

# Core Agent Framework
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107

# Agent Server (HTTP hosting)
azure-ai-agentserver-agentframework==1.0.0b10
azure-ai-agentserver-core==1.0.0b10

# Azure Identity (Managed Identity / DefaultAzureCredential)
azure-identity>=1.19.0

# OpenTelemetry + Azure Monitor
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0

# HTTP client (para test_agent.py - console client)
requests>=2.31.0

# Azure AI Projects SDK (para create_hosted_agent.py)
azure-ai-projects>=2.0.0b3

# Utilitarios
python-dotenv>=1.0.0
